requirements.txt
absl-py==2.2.2
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiohttp-cors==0.8.1
aiosignal==1.3.2
airportsdata==20250224
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
astor==0.8.1
attrs==25.3.0
av==14.2.0
blake3==1.0.4
boto3==1.37.28
botocore==1.37.28
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
codetiming==1.4.0
colorful==0.5.6
compressed-tensors==0.9.2
cupy-cuda12x==13.4.1
datasets==3.5.0
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
docker-pycreds==0.4.0
einops==0.8.1
email-validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
fastrlock==0.8.3
filelock==3.18.0
# flash-attn==2.7.3
frozenlist==1.5.0
fsspec==2024.12.0
gguf==0.10.0
gitdb==4.0.12
gitpython==3.1.44
google-api-core==2.24.2
google-auth==2.38.0
googleapis-common-protos==1.69.2
grpcio==1.71.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.1
idna==3.10
importlib-metadata==8.6.1
iniconfig==2.1.0
interegular==0.3.3
jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lark==1.2.2
liger-kernel==0.5.6
linkify-it-py==2.0.3
llvmlite==0.43.0
lm-format-enforcer==0.10.11
markdown==3.7
markdown-it-py==3.0.0
markupsafe==3.0.2
mathruler==0.1.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
memray==1.17.0
mistral-common==1.5.4
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
# multidict==6.3.2
multidict
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.4.2
# numba==0.60.0
numpy==2.2.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
nvitop==1.4.2
omegaconf==2.3.0
openai==1.70.0
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python-headless==4.11.0.86
orjson==3.10.16
outlines==0.1.11
outlines-core==0.1.26
packaging==24.2
pandas==2.2.3
partial-json-parser==0.2.1.1.post5
peft==0.15.1
pillow==11.1.0
platformdirs==4.3.7
pluggy==1.5.0
prometheus-client==0.21.1
prometheus-fastapi-instrumentator==7.1.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
py-cpuinfo==9.0.0
py-spy==0.4.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pybind11==2.13.6
pycountry==24.6.1
pydantic==2.11.2
pydantic-core==2.33.1
pygments==2.19.1
pylatexenc==2.10
pynvml==12.0.0
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
pyzmq==26.4.0
qwen-vl-utils==0.0.10
ray==2.40.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.1
rpds-py==0.24.0
rsa==4.9
s3transfer==0.11.4
safetensors==0.5.3
sentencepiece==0.2.0
sentry-sdk==2.25.1
setproctitle==1.3.5
setuptools==78.1.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.1
swankit==0.1.6
swanlab==0.5.4
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensordict==0.6.2
textual==3.0.1
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.5.1
torchaudio==2.5.1
torchdata==0.11.0
torchvision==0.20.1
tqdm==4.67.1
transformers==4.50.3
triton==3.1.0
typer==0.15.2
typing-extensions==4.13.1
typing-inspection==0.4.0
tzdata==2025.2
uc-micro-py==1.0.3
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.30.0
# vllm==0.7.4.dev65+g22757848
wandb==0.19.9
watchfiles==1.0.4
websockets==15.0.1
werkzeug==3.1.3
wrapt==1.17.2
xformers==0.0.28.post3
xgrammar==0.1.11
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
